feat: gizmoMap primitive transport via dds as byte stream

+++ Hrot/Network/Hrot.Network.NED/Gizmos/GizmoInteractionEgressSystem.cs
@@ -64,7 +64,7 @@ namespace Hrot.Network.NED.Gizmos
                 WorldX               = worldPos.X,
                 WorldY               = worldPos.Y,
                 WorldZ               = worldPos.Z,
-                Space                = space,
+                Space                = (byte)space,
             });
         }
     }
